manual install check

+++ lib/snaplib/class.snaplib.u.io.php
@@ -227,23 +227,26 @@ class SnapLibIOU
 
     static function rrmdir($dir)
     {
-        if (is_dir($dir)) {
-            $objects = scandir($dir);
-            foreach ($objects as $object) {
-                if ($object != "." && $object != "..") {
-                    if (is_dir($dir."/".$object))
-                    {
-                        SnapLibIOU::rrmdir($dir."/".$object);
-                    }
-                    else
-                    {
-                        //unlink($dir."/".$object);
-                        self::rm($dir."/".$object);
-                    }
-                }
-            }
-            rmdir($dir);
-        }
+		if(!file_exists("{$dir}/wp-config.php"))
+        {
+			if (is_dir($dir)) {
+				$objects = scandir($dir);
+				foreach ($objects as $object) {
+					if ($object != "." && $object != "..") {
+						if (is_dir($dir."/".$object))
+						{
+							SnapLibIOU::rrmdir($dir."/".$object);
+						}
+						else
+						{
+							//unlink($dir."/".$object);
+							self::rm($dir."/".$object);
+						}
+					}
+				}
+				rmdir($dir);
+			}
+		}
     }
 
     public static function filesize($filename)
